+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- #288 Add Balance Get method - @rbutler
 - #286 Deserialize meta - @timoreimann
 
 ## [v1.28.0] - 2019-12-04